requirements.txt
pyrogram
tgcrypto
ffmpeg-python
pillow
rarfile
zipfile
